test_scissor
